refactor: move stmt group to another file
