+++ backend/plugin/schema/mysql/get_design_schema.go
@@ -22,21 +22,6 @@ func init() {
 	schema.RegisterGetDesignSchema(storepb.Engine_MYSQL, GetDesignSchema)
 }
 
-type groupType int
-
-const (
-	groupTypeUnknown         groupType = iota
-	groupTypeCreateFunction            // CREATE FUNCTION
-	groupTypeCreateProcedure           // CREATE PROCEDURE
-)
-
-type stmtGroup struct {
-	tp         groupType
-	beginIdx   int
-	endIdx     int
-	objectName string
-}
-
 func GetDesignSchema(_, baselineSchema string, to *storepb.DatabaseSchemaMetadata) (string, error) {
 	toState := convertToDatabaseState(to)
 	list, err := mysqlparser.ParseMySQL(baselineSchema, tokenizer.KeepEmptyBlocks())
@@ -62,71 +47,9 @@ func GetDesignSchema(_, baselineSchema string, to *storepb.DatabaseSchemaMetadat
 	listener.finalViewTail = finalViewTailGetter.finalViewTails
 	listener.finalViewStatementIndex = finalViewTailGetter.finalViewStatementIndex
 
-	infoGetter := &stmtInfoGetter{}
-	for _, stmt := range list {
-		antlr.ParseTreeWalkerDefault.Walk(infoGetter, stmt.Tree)
-	}
-	infos := infoGetter.result
-
-	// Constructs the types into groups.
-	var groups []stmtGroup
-	for i := 0; i < len(infos); i++ {
-		nextIdx := i
-		if infos[i].tp == currentStmtTypeSet {
-			// Find the next SET/CREATE FUNCTION/CREATE PROCEDURE statement.
-			j := i + 1
-			for ; j < len(infos); j++ {
-				if infos[j].tp == currentStmtTypeSet {
-					continue
-				}
-				if infos[j].tp == currentStmtTypeCreateFunction {
-					nextIdx = j + 1
-					groups = append(groups, stmtGroup{
-						tp:         groupTypeCreateFunction,
-						beginIdx:   i,
-						endIdx:     j,
-						objectName: infos[j].normalizedObjectName,
-					})
-					break
-				}
-				if infos[j].tp == currentStmtTypeCreateProcedure {
-					nextIdx = j + 1
-					groups = append(groups, stmtGroup{
-						tp:         groupTypeCreateProcedure,
-						beginIdx:   i,
-						endIdx:     j,
-						objectName: infos[j].normalizedObjectName,
-					})
-					break
-				}
-				break
-			}
-		}
-		if infos[i].tp == currentStmtTypeCreateFunction {
-			groups = append(groups, stmtGroup{
-				tp:         groupTypeCreateFunction,
-				beginIdx:   i,
-				endIdx:     i,
-				objectName: infos[i].normalizedObjectName,
-			})
-		}
-		if infos[i].tp == currentStmtTypeCreateProcedure {
-			groups = append(groups, stmtGroup{
-				tp:         groupTypeCreateProcedure,
-				beginIdx:   i,
-				endIdx:     i,
-				objectName: infos[i].normalizedObjectName,
-			})
-		}
-		if nextIdx > i {
-			i = nextIdx - 1
-		} else {
-			groups = append(groups, stmtGroup{
-				tp:       groupTypeUnknown,
-				beginIdx: i,
-				endIdx:   i,
-			})
-		}
+	groups, err := groupStatement(list)
+	if err != nil {
+		return "", err
 	}
 
 	groupIdx := 0
@@ -223,88 +146,6 @@ func GetDesignSchema(_, baselineSchema string, to *storepb.DatabaseSchemaMetadat
 		return s, nil
 	}
 	return result, nil
-}
-
-type stmtType int
-
-const (
-	currentStmtTypeUnknown stmtType = iota
-	currentStmtTypeCreateFunction
-	currentStmtTypeCreateProcedure
-	currentStmtTypeSet
-)
-
-type stmtInfo struct {
-	tp                   stmtType
-	normalizedObjectName string
-}
-
-// stmtInfoGetter is used to get the type of the current statement.
-// The group logic described as follows:
-// 1. If the statement is CREATE FUNCTION, and the previous continuous SET statement would be grouped into the createFunctionGroup.
-// 2. If the statement is CREATE PROCEDURE, and the previous continuous SET statement would be grouped into the createProcedureGroup.
-type stmtInfoGetter struct {
-	*mysql.BaseMySQLParserListener
-
-	tp                   stmtType
-	normalizedObjectName string
-
-	result []*stmtInfo
-}
-
-func (g *stmtInfoGetter) ExitQuery(*mysql.QueryContext) {
-	g.result = append(g.result, &stmtInfo{
-		tp:                   g.tp,
-		normalizedObjectName: g.normalizedObjectName,
-	})
-	g.tp = currentStmtTypeUnknown
-	g.normalizedObjectName = ""
-}
-
-func (g *stmtInfoGetter) EnterSetStatement(ctx *mysql.SetStatementContext) {
-	p := ctx.GetParent()
-	if _, ok := p.(*mysql.SimpleStatementContext); !ok {
-		return
-	}
-	pp := p.GetParent()
-	if _, ok := pp.(*mysql.QueryContext); !ok {
-		return
-	}
-	g.tp = currentStmtTypeSet
-}
-
-func (g *stmtInfoGetter) EnterCreateProcedure(ctx *mysql.CreateProcedureContext) {
-	p := ctx.GetParent()
-	if _, ok := p.(*mysql.CreateStatementContext); !ok {
-		return
-	}
-	pp := p.GetParent()
-	if _, ok := pp.(*mysql.SimpleStatementContext); !ok {
-		return
-	}
-	ppp := pp.GetParent()
-	if _, ok := ppp.(*mysql.QueryContext); !ok {
-		return
-	}
-	g.tp = currentStmtTypeCreateProcedure
-	_, g.normalizedObjectName = mysqlparser.NormalizeMySQLProcedureName(ctx.ProcedureName())
-}
-
-func (g *stmtInfoGetter) EnterCreateFunction(ctx *mysql.CreateFunctionContext) {
-	p := ctx.GetParent()
-	if _, ok := p.(*mysql.CreateStatementContext); !ok {
-		return
-	}
-	pp := p.GetParent()
-	if _, ok := pp.(*mysql.SimpleStatementContext); !ok {
-		return
-	}
-	ppp := pp.GetParent()
-	if _, ok := ppp.(*mysql.QueryContext); !ok {
-		return
-	}
-	g.tp = currentStmtTypeCreateFunction
-	_, g.normalizedObjectName = mysqlparser.NormalizeMySQLFunctionName(ctx.FunctionName())
 }
 
 type finalViewTailGetter struct {
